+++ net.tourbook/src/net/tourbook/srtm/ElevationSRTM3.java
@@ -56,7 +56,7 @@ public final class ElevationSRTM3 extends ElevationBase {
 			}
 		}
 
-		private short getElevation(final GeoLat lat, final GeoLon lon) {
+		private float getElevation(final GeoLat lat, final GeoLon lon) {
 			return elevationFile.get(srtmFileOffset(lat, lon));
 		}
 
@@ -171,7 +171,7 @@ public final class ElevationSRTM3 extends ElevationBase {
 	}
 
 	@Override
-	public short getElevation(final GeoLat lat, final GeoLon lon) {
+	public float getElevation(final GeoLat lat, final GeoLon lon) {
 
 		if (lat.tertias != 0) {
 			return getElevationGrid(lat, lon);
